requirements.txt
esptool
setuptools
pyserial
mpfshell
adafruit-ampy
